修改部分数据类型

+++ rap-data-common/src/main/java/com/ifugle/rap/utils/UserOds.java
@@ -84,7 +84,7 @@ public class UserOds implements Serializable {
      * 人员类型代码(1: 法人 2: 财务负责人 3: 办税人 4: 其他办税人 5: 购票员)
      */
     @SerializedName("rylx_dm")
-    private Byte rylxDm;
+    private Integer rylxDm;
 
     /**
      * 认证级别(0=未认证，1=强制认证，2=短信认证，16=刷脸认证)
